+++ Front-End/src/app/appCtrl.js
@@ -1,0 +1,10 @@
+'use strict';
+
+// Declare app level module which depends on views, and components
+angular.controller('appCtrl', [
+    '$http'
+    , function ($locationProvider, $routeProvider) {
+        $locationProvider.hashPrefix('!');
+
+        $routeProvider.otherwise({redirectTo: '/view1'});
+    }]);